move entrypoint and view to books package in wishlist-frontend and remove hello_reagent_reframe.core
Until now I still had this code running at the same time, which was actually from a tutorial. The idea was to be able to easily look up this tutorial, since it was in the same code. But as of now this code gives me no more hints than my own code already does, so now I am deleting it.

Note that the todomvc code examples are still in the repo, as these may well turn out to useful still!