remove files in mongo folder: quick_actions.clj and books_from_wishlist_faulty.clj
Both of these files are quite old and I haven't looked into them for quite some time. I am sure that both hint to some real errors when importing from Amazon.

However, I'm also sure that I will find plenty errors there, for that matter. As I'm currently not working on that, I will see when I work again on that, and then check for errors on the go...

#BOOKS-22