in amazon.books.fetch.scraping.driver, change defs to defns, as they use env and cannot be def'd therefore, as mount has not yet started in this case
Once mount starts, it will initialize env and its environment variables. In driver.clj, these def's depended on those. However, when use def, Clojure will try to statically solve for these, which is not possible due to mount, which is only started on runtime. We therefore changed all to a defn and it now works again.

#BOOKS-22